Add a nightly version that is known to work
Dearest Reviewer,

I have had issues with building on nightly and some dependencies. I got
this version of rust to work. In the past some dependencies seem to not
be compiling on nightly. If there is a better version of nightly let me
know and I will change it. Or just close the PR.

Thanks! I look forward to using it.
Becker